requirements.txt
chardet==3.0.4
pdfminer.six==20200124
pycryptodome==3.9.6
sortedcontainers==2.1.0